\name{BFInfo}
\alias{BFInfo}

\title{
Prints the version information for the BayesFactor package
}
\description{
Prints the version, revision, and date information for the BayesFactor package

}
\usage{
BFInfo()
}

\details{
This function prints the version, revision, and date information for the BayesFactor package. It takes no arguments, and returns nothing.
}
\value{
This function returns nothing.
}
\author{
Richard D. Morey
}

\keyword{misc}