\name{puzzles}
\alias{puzzles}
\docType{data}
\title{Puzzle completion times from Hays (1994)}
\description{
   DESCRIPTION
}
\usage{data(puzzles)}
\format{
  A data frame with 48 observations on 3 variables.
  \describe{
  	\item{\code{RT}}{Response time}
    \item{\code{ID}}{the subject identifier}
    \item{\code{shape}}{shape of the puzzle (round or square)}
    \item{\code{color}}{color content of the puzzle (monochromatic or color)}
  }
}
\details{
	Hays (section 13.21, table 13.21.2, p. 570)
}
\source{
Hays, W. L. (1994), Statistics (5th edition), Harcourt Brace, Fort Worth, Texas
\examples{
data(puzzles)

## ANOVA

## Bayes Factor

}
\keyword{datasets}
